refactor(eval): extract load_custom_prompts to shared _anthropic_api module
Consolidate duplicate load_custom_prompts function that existed in both
eval-agents.py and eval-knowledge-integration.py into the shared
_anthropic_api.py module where other common utilities already reside.

This eliminates code duplication and prevents silent divergence if the
JSON parsing logic needs to change in the future.
