+++ scripts/eval/eval-knowledge-integration.py
@@ -25,7 +25,7 @@ from typing import Any
 # API utilities (shared module)
 # ---------------------------------------------------------------------------
 
-from _anthropic_api import call_api as _call_api, load_api_key as _load_api_key
+from _anthropic_api import call_api as _call_api, load_api_key as _load_api_key, load_custom_prompts
 
 
 # ---------------------------------------------------------------------------
@@ -331,26 +331,6 @@ def run_assessment(
         }
 
     return results
-
-
-def load_custom_prompts(path: str) -> dict[str, list[dict[str, str]]]:
-    """Load prompts from a JSON file.
-
-    Expected format:
-    {
-        "skill-name": [
-            {"prompt": "...", "expected": "..."},
-            ...
-        ]
-    }
-    """
-    with open(path) as f:
-        data = json.load(f)
-
-    # Support both {"prompts": {...}} wrapper and direct dict
-    if "prompts" in data and isinstance(data["prompts"], dict):
-        return data["prompts"]
-    return data
 
 
 def main() -> None:
